Review feedback

+++ nexus/db-model/src/lib.rs
@@ -64,7 +64,7 @@ mod silo_user_password_hash;
 mod sled;
 mod snapshot;
 mod ssh_key;
-mod u16;
+mod unsigned;
 mod update_artifact;
 mod user_builtin;
 mod virtual_provisioning_collection;
@@ -87,7 +87,7 @@ mod db {
 }
 
 pub use self::macaddr::*;
-pub use self::u16::*;
+pub use self::unsigned::*;
 pub use block_size::*;
 pub use bytecount::*;
 pub use certificate::*;
